hotfix/Circular Module Dependency on generateHash function
