+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@mojaloop/central-services-shared",
-  "version": "7.4.2",
+  "version": "7.4.3",
   "description": "Shared code for central services",
   "main": "src/index.js",
   "scripts": {
