+++ src/util/hash.js
@@ -3,11 +3,8 @@
  --------------
  Copyright © 2017 Bill & Melinda Gates Foundation
  The Mojaloop files are made available by the Bill & Melinda Gates Foundation under the Apache License, Version 2.0 (the "License") and you may not use these files except in compliance with the License. You may obtain a copy of the License at
-
  http://www.apache.org/licenses/LICENSE-2.0
-
  Unless required by applicable law or agreed to in writing, the Mojaloop files are distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-
  Contributors
  --------------
  This is the official list of the Mojaloop project contributors for this file.
@@ -18,45 +15,26 @@
  Gates Foundation organization for an example). Those individuals should have
  their names indented and be marked with a '-'. Email address can be added
  optionally within square brackets <email>.
-
  * Gates Foundation
- - Name Surname <name.surname@gatesfoundation.com>
+ * Name Surname <name.surname@gatesfoundation.com>
 
  * Georgi Georgiev <georgi.georgiev@modusbox.com>
- * Miguel de Barros <miguel.debarros@modusbox.com>
+ * Rajiv Mothilal <rajiv.mothilal@modusbox.com>
  --------------
  ******/
 'use strict'
 
-const Hash = require('../hash')
+const Crypto = require('crypto')
 
-const duplicateCheckComparator = async (id, object, getDuplicateDataFuncOverride, saveHashFuncOverride) => {
-  let hasDuplicateId = false
-  let hasDuplicateHash = false
-  const generatedHash = Hash.generateSha256(object)
-  let duplicateHashRecord
-
-  const compareById = async () => {
-    duplicateHashRecord = await getDuplicateDataFuncOverride(id)
-    hasDuplicateId = duplicateHashRecord !== null
-    return hasDuplicateId
-  }
-
-  const compareByHash = () => {
-    hasDuplicateHash = duplicateHashRecord.hash === generatedHash
-    return hasDuplicateHash
-  }
-
-  if (await compareById()) {
-    compareByHash()
-  } else {
-    await saveHashFuncOverride(id, generatedHash)
-  }
-
-  return {
-    hasDuplicateId,
-    hasDuplicateHash
-  }
+function generateSha256 (object) {
+  const hashSha256 = Crypto.createHash('sha256')
+  let hash = JSON.stringify(object)
+  hash = hashSha256.update(hash)
+  // remove trailing '=' as per specification
+  hash = hashSha256.digest(hash).toString('base64').slice(0, -1)
+  return hash
 }
 
-module.exports = duplicateCheckComparator
+module.exports = {
+  generateSha256
+}
