+++ src/util/index.js
@@ -26,7 +26,6 @@
 'use strict'
 
 const _ = require('lodash')
-const Crypto = require('crypto')
 const Kafka = require('./kafka')
 const Endpoints = require('./endpoints')
 const Request = require('./request')
@@ -36,6 +35,7 @@ const Headers = require('./headers/transformer')
 const Encoding = require('./encoding')
 const StreamingProtocol = require('./streaming/protocol')
 const Time = require('./time')
+const Hash = require('./hash')
 const Comparators = require('./comparators/index')
 
 const omitNil = (object) => {
@@ -185,15 +185,6 @@ const transpose = function (obj) {
   return transposed
 }
 
-const generateHash = (object) => {
-  const hashSha256 = Crypto.createHash('sha256')
-  let hash = JSON.stringify(object)
-  hash = hashSha256.update(hash)
-  // remove trailing '=' as per specification
-  hash = hashSha256.digest(hash).toString('base64').slice(0, -1)
-  return hash
-}
-
 module.exports = {
   assign,
   expand,
@@ -211,7 +202,6 @@ module.exports = {
   setValueByCaseInsensitiveKey,
   breadcrumb,
   transpose,
-  generateHash,
   Kafka,
   Endpoints,
   Request,
@@ -221,5 +211,6 @@ module.exports = {
   Encoding,
   StreamingProtocol,
   Time,
+  Hash,
   Comparators
 }
