+++ test/unit/util/hash.test.js
@@ -3,11 +3,8 @@
  --------------
  Copyright © 2017 Bill & Melinda Gates Foundation
  The Mojaloop files are made available by the Bill & Melinda Gates Foundation under the Apache License, Version 2.0 (the "License") and you may not use these files except in compliance with the License. You may obtain a copy of the License at
-
  http://www.apache.org/licenses/LICENSE-2.0
-
  Unless required by applicable law or agreed to in writing, the Mojaloop files are distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-
  Contributors
  --------------
  This is the official list of the Mojaloop project contributors for this file.
@@ -18,45 +15,39 @@
  Gates Foundation organization for an example). Those individuals should have
  their names indented and be marked with a '-'. Email address can be added
  optionally within square brackets <email>.
-
  * Gates Foundation
  - Name Surname <name.surname@gatesfoundation.com>
 
  * Georgi Georgiev <georgi.georgiev@modusbox.com>
- * Miguel de Barros <miguel.debarros@modusbox.com>
  --------------
  ******/
+
 'use strict'
 
-const Hash = require('../hash')
+const Test = require('tape')
+const Hash = require('../../../src/util/hash')
 
-const duplicateCheckComparator = async (id, object, getDuplicateDataFuncOverride, saveHashFuncOverride) => {
-  let hasDuplicateId = false
-  let hasDuplicateHash = false
-  const generatedHash = Hash.generateSha256(object)
-  let duplicateHashRecord
+Test('Hash util', hashTest => {
+  hashTest.test('generateSha256 should', generateSha256Test => {
+    generateSha256Test.test('return hash for a given object', test => {
+      const obj1 = {
+        prop1: 'test',
+        prop2: {
+          date_time: 'Thu Aug 22 2019"',
+          number: 1000
+        },
+        prop3: null,
+        prop4: null
+      }
 
-  const compareById = async () => {
-    duplicateHashRecord = await getDuplicateDataFuncOverride(id)
-    hasDuplicateId = duplicateHashRecord !== null
-    return hasDuplicateId
-  }
+      const expected = 'NLI18feAfVIGLKjz8Lq5VH66t89cVgkYDMedHBEev88'
 
-  const compareByHash = () => {
-    hasDuplicateHash = duplicateHashRecord.hash === generatedHash
-    return hasDuplicateHash
-  }
+      const result = Hash.generateSha256(obj1)
+      test.equal(result, expected)
+      test.end()
+    })
+    generateSha256Test.end()
+  })
 
-  if (await compareById()) {
-    compareByHash()
-  } else {
-    await saveHashFuncOverride(id, generatedHash)
-  }
-
-  return {
-    hasDuplicateId,
-    hasDuplicateHash
-  }
-}
-
-module.exports = duplicateCheckComparator
+  hashTest.end()
+})
