+++ test/unit/util/index.test.js
@@ -552,26 +552,5 @@ Test('General util', utilTest => {
     transposeTest.end()
   })
 
-  utilTest.test('generateHash should', generateHashTest => {
-    generateHashTest.test('return hash for a given object', test => {
-      const obj1 = {
-        prop1: 'test',
-        prop2: {
-          date_time: 'Thu Aug 22 2019"',
-          number: 1000
-        },
-        prop3: null,
-        prop4: null
-      }
-
-      const expected = 'NLI18feAfVIGLKjz8Lq5VH66t89cVgkYDMedHBEev88'
-
-      const result = Util.generateHash(obj1)
-      test.equal(result, expected)
-      test.end()
-    })
-    generateHashTest.end()
-  })
-
   utilTest.end()
 })
